+++ src/components/Modify.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react'
 import styled from "styled-components"
 import { useSelector } from "react-redux"
 import { NormalButton, ButtonContainer } from './Main'
+import Swal from "sweetalert2"
 
 const Form = styled.form`
   display: flex;
@@ -67,17 +68,52 @@ interface IsRender {
     if(state.index){
       //delete는 현재 노트값 삭제
       if(e.target.value === 'del'){
-        dataBase.splice(state.index, 1)
+        Swal.fire({
+          title: '정말로 이 노트를 삭제할까요?',
+          text: "한번 삭제하면 되돌릴 수 없어용!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#e99797',
+          cancelButtonColor: '#4ec6e4',
+          confirmButtonText: '그래! 지워버려~',
+          cancelButtonText: '시러~'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire({
+              icon: 'success',
+              title: '삭제완료',
+              html: `
+              너의 노트. 아주 깔끔하게 사라졌다.
+              `,
+              showConfirmButton: false,
+              timer: 1000
+            })
+
+            dataBase.splice(state.index, 1)
+            const jsonDB = JSON.stringify(dataBase)
+            localStorage.setItem(`memoKey`, jsonDB)
+            props.isRender()
+            return
+          }else{
+            return
+          }
+        })
       }else{
       //수정 후 데이터베이스 추가 
         dataBase.splice(state.index, 1, currentDB)
+        Swal.fire({
+          icon: 'success',
+          title: '노트 수정 완료!',
+          showConfirmButton: false,
+          timer: 800
+        })
+        const jsonDB = JSON.stringify(dataBase)
+        localStorage.setItem(`memoKey`, jsonDB)
+        props.isRender()
+        return
       }
 
-      const jsonDB = JSON.stringify(dataBase)
-      localStorage.setItem(`memoKey`, jsonDB)
-      props.isRender()
-
-    //새노트 추가일 경우
+    //새 노트 추가일 경우
     }else if(title !== state.title && contents !== state.contents){
       //delete는 현재 작성내용을 초기화시키는 역할부여
       if(e.target.value === 'del'){
@@ -91,14 +127,34 @@ interface IsRender {
       const jsonDB = JSON.stringify(dataBase)
       localStorage.setItem(`memoKey`, jsonDB)
       props.isRender()
+      Swal.fire({
+        icon: 'success',
+        title: '노트 추가 성공!',
+        html: `
+        새 노트가 추가되었어요!
+        `,
+        showConfirmButton: false,
+        timer: 1000
+      })
+      return
+    }
+    if(e.target.value !== 'del'){
+      Swal.fire({
+        html: `
+        제목과 내용을 모두 입력해주세요!
+        `,
+        icon: 'warning',
+        showCancelButton: false,
+        confirmButtonText: "알겠습니다...",
+      })
     }
   }
 
   return (
     <>
       <Form onSubmit={handlerFormSubmit}>
-        <TitleArea type="text" value={title} onChange={(e)=> setTitle(e.target.value)}></TitleArea>
-        <ContentArea value={contents} onChange={(e) => setContents(e.target.value)}></ContentArea>
+        <TitleArea type="text" value={title} onChange={(e)=> setTitle(e.target.value)} placeholder="제목을 입력해주세요..."></TitleArea>
+        <ContentArea value={contents} onChange={(e) => setContents(e.target.value)} placeholder="내용을 입력해주세요..."></ContentArea>
         <NormalButton>Done</NormalButton>
       </Form>
       <ButtonContainer>
